Refactor Sleep stage breakdown UI
